cloud-functions/runtask-process/requirements.txt
functions-framework==3.4.0
flask==2.3.2
google-cloud-error-reporting==1.9.2
google-cloud-logging==3.6.0
